Move health services
Also rename `service-status.service.js` as `info.service.js`.
