+++ app/services/health/info.service.js
@@ -1,8 +1,8 @@
 'use strict'
 
 /**
- * Checks status, including versions, of the various services which make up WRLS
- * @module ServiceStatusService
+ * Checks status and gathers info for each of the services which make up WRLS
+ * @module InfoService
  */
 
 // We use promisify to wrap exec in a promise. This allows us to await it without resorting to using callbacks.
@@ -10,7 +10,7 @@ const ChildProcess = require('child_process')
 const util = require('util')
 const exec = util.promisify(ChildProcess.exec)
 
-const servicesConfig = require('../../config/services.config.js')
+const servicesConfig = require('../../../config/services.config.js')
 
 /**
  * Returns data required to populate our `/service-status` page, eg. task activity status, virus checker status, service
